+++ src_c/bitmask.c
@@ -20,7 +20,7 @@
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "bitmask.h"
+#include "../include/bitmask.h"
 #include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
